+++ vck-openid-ktor/src/commonMain/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClient.kt
@@ -7,6 +7,8 @@ import at.asitplus.openid.*
 import at.asitplus.openid.OpenIdConstants.AUTH_METHOD_ATTEST_JWT_CLIENT_AUTH
 import at.asitplus.openid.OpenIdConstants.PARAMETER_PROMPT
 import at.asitplus.openid.OpenIdConstants.PARAMETER_PROMPT_LOGIN
+import at.asitplus.openid.OpenIdConstants.PATH_WELL_KNOWN_OAUTH_AUTHORIZATION_SERVER
+import at.asitplus.openid.OpenIdConstants.PATH_WELL_KNOWN_OPENID_CONFIGURATION
 import at.asitplus.openid.OpenIdConstants.TOKEN_TYPE_DPOP
 import at.asitplus.wallet.lib.agent.DefaultCryptoService
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
@@ -166,9 +168,13 @@ class OpenId4VciClient(
         val issuerMetadata = credentialIdentifierInfo.issuerMetadata
         val authorizationServer = issuerMetadata.authorizationServers?.firstOrNull()
             ?: credentialIssuerUrl
-        val oauthMetadata = client
-            .get("$authorizationServer${OpenIdConstants.PATH_WELL_KNOWN_OPENID_CONFIGURATION}")
-            .body<OAuth2AuthorizationServerMetadata>()
+        val oauthMetadata = catching {
+            client.get("$authorizationServer$PATH_WELL_KNOWN_OAUTH_AUTHORIZATION_SERVER")
+                .body<OAuth2AuthorizationServerMetadata>()
+        }.getOrElse {
+            client.get("$authorizationServer$PATH_WELL_KNOWN_OPENID_CONFIGURATION")
+                .body<OAuth2AuthorizationServerMetadata>()
+        }
 
         val state = uuid4().toString()
         ProvisioningContext(
@@ -442,9 +448,13 @@ class OpenId4VciClient(
         val issuerMetadata = credentialIdentifierInfo.issuerMetadata
         val authorizationServer = issuerMetadata.authorizationServers?.firstOrNull()
             ?: credentialOffer.credentialIssuer
-        val oauthMetadata = client
-            .get("$authorizationServer${OpenIdConstants.PATH_WELL_KNOWN_OPENID_CONFIGURATION}")
-            .body<OAuth2AuthorizationServerMetadata>()
+        val oauthMetadata = catching {
+            client.get("$authorizationServer$PATH_WELL_KNOWN_OAUTH_AUTHORIZATION_SERVER")
+                .body<OAuth2AuthorizationServerMetadata>()
+        }.getOrElse {
+            client.get("$authorizationServer$PATH_WELL_KNOWN_OPENID_CONFIGURATION")
+                .body<OAuth2AuthorizationServerMetadata>()
+        }
         val state = uuid4().toString()
 
         credentialOffer.grants?.preAuthorizedCode?.let {